Get wgpu rendering working
There's a lot going on in this commit, but the only part that
matters is this:

We changed the color storage buffer from an array of vec3 to
an array of vec4, and changed the color buffer contents to include
the color's alpha (always 1.0).

The problem was that the vertex shader was always indexing the
color storage buffer as if it had a stride of 16 (4 f32s) even
though it was clearly typed as being an array of vec3f (stride of 12)
sooooo to fix it just put elements of 16 bytes in the buffer and
call it a day!