Fix the color struct so that the color buffer works
storage buffers in webgpu have to have alignments of 4, 8, or 16.
Even though the buffer was an array(vec3f), it indexed it using
the alignment for an array(vec4f).

The fix is to give the Color struct a #[repr(C, align(16))], so
that when turning it into byts for the buffer it pads the last 4
bytes.

Removes the alpha component of the color just to satisfy me.